feat:configuração do concierge de investimentos, cotação de dólar e compra de ações

+++ models/cliente.py
@@ -8,6 +8,6 @@ class Cliente(db.Model):
     nome = db.Column(db.String(100), nullable=False,)
     agencia = db.Column(db.String(10), nullable=False)
     conta = db.Column(db.String(20), nullable=False)
-    saldo = db.Column(db.Float, nullable=False)
+    saldo = db.Column(db.Numeric(10,2), nullable=False)
     nivel = db.Column(db.String(20), nullable=False)
     produtos = db.Column(db.String(200), nullable=True)
